퇴장 시 스코어 보드 즉각 반영 되게끔 수정 / 퇴장 과정 논리적으로 수정 / authenticationEntryPoint오류 처리 / 관리자페이지 접근 시 접근 권한 필요 하다는 페이지 생성 / 로그인 한 유저가 재 로그인 시도하는 경우 홈으로 redirect되게끔 CustomLoginPageFilter 생성 및 적용

+++ guess_music/src/main/java/com/example/guess_music/controller/MessageController.java
@@ -52,9 +52,11 @@ public class MessageController {
             }
         }
         if(ChatMessage.MessageType.LEAVE.equals(message.getType())){
-            //유저 퇴장 시 해당 User객체 삭제
-            gameService.deleteUserByUsername(message.getSender());
-            //및 게임 방의 유저 수 --
+            if(room.getRoomStatus().equals("START")){
+                //유저 퇴장 시 해당 User객체 삭제
+                gameService.deleteUserByUsername(message.getSender());
+            }
+            //게임 방의 유저 수 --
             room.setRoomUserNum(room.getRoomUserNum() - 1);
             if(room.getRoomUserNum() == 0){
                 //아무도 없으면 방 삭제
